+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- do not terminate watcher after failures.
 - add the commit-log details to the Component object, so then it'll be available for `bit show --json` and `bit export`.
 
 ## [0.6.0] - 2017-05-15